the two new activities were put in the wrong folder, Moved them to the right place
