+++ Android/app/src/main/java/com/example/android/Views/Marketplace_PlayerInventory_Startup.java
@@ -1,4 +1,4 @@
-package com.example.android;
+package com.example.android.Views;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -6,6 +6,8 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
+
+import com.example.android.R;
 
 public class Marketplace_PlayerInventory_Startup extends AppCompatActivity {
