+++ Android/app/src/main/res/layout/activity_marketplace__player_inventory__startup.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Marketplace_PlayerInventory_Startup">
+    tools:context=".Views.Marketplace_PlayerInventory_Startup">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
